refactor: finish AOT review follow-up
Restore the missing extracted AOT metadata, keep the legacy constants, and move the remaining MojoUtils implementation into micronaut-maven-core while leaving the plugin class as a compatibility facade.
